+++ src/components/TransactionsItem/TransactionsItem.jsx
@@ -1,3 +1,4 @@
+import ModalEditTransaction from "../ModalEditTransaction/ModalEditTransaction";
 import s from "./TransactionsItem.module.css";
 
 const TransactionsItem = ({ transaction }) => {
@@ -11,7 +12,7 @@ const TransactionsItem = ({ transaction }) => {
         {transaction.sum.toFixed(2)}
       </td>
       <td className={s.actionBtn}>
-        <button className={s.editBtn}>
+        <button onClick={ModalEditTransaction} className={s.editBtn}>
           <svg className={s.editIcon} width="12" height="12" stroke="#fafafa">
             <use href="../../../src/images/symbol-defs.svg#pencil"></use>
           </svg>
